refactor: Change Dataset to follow folder structure convention
