+++ core/src/datasets/index.ts
@@ -1,0 +1,1 @@
+export { Dataset, DatasetProps } from './dataset';
