+++ core/src/index.ts
@@ -4,7 +4,7 @@
 export { ExampleProps, Example } from './example';
 export { DataLakeStorageProps, DataLakeStorage } from './data-lake-storage';
 export { SynchronousCrawlerProps, SynchronousCrawler } from './synchronous-crawler';
-export { DatasetProps, Dataset } from './dataset';
+export { DatasetProps, Dataset } from './datasets';
 export { DataGeneratorProps, DataGenerator } from './data-generator';
 export { SynchronousAthenaQueryProps, SynchronousAthenaQuery } from './synchronous-athena-query';
 export { SingletonBucket } from './singleton-bucket';
