+++ core/test/data-generator.test.ts
@@ -3,7 +3,7 @@
 
 import { Stack } from '@aws-cdk/core';
 import { DataGenerator } from '../src/data-generator';
-import { Dataset } from '../src/dataset';
+import { Dataset } from '../src/datasets';
 import '@aws-cdk/assert/jest';
 
 const dataGeneratorStack = new Stack();
