+++ core/test/datasets/dataset.test.ts
@@ -1,8 +1,8 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: MIT-0
 
-import { DataGenerator } from '../src/data-generator';
-import { Dataset } from '../src/dataset';
+import { DataGenerator } from '../../src/data-generator';
+import { Dataset } from '../../src/datasets';
 import '@aws-cdk/assert/jest';
